Fix incorrect memcpy usage
Identified an issue with double pointer addresses being used.
Changed size parameter to explicitly casted size_t
Pulled apart function chain for temporary debugging

+++ src/messages/TemperatureMessage.h
@@ -9,14 +9,15 @@
 #define TEMPERATUREMESSAGE_H_
 
 #include <src/messages/Message.h>
+#include <stdint.h>
 
 class TemperatureMessage: public Message {
 public:
     TemperatureMessage();
     TemperatureMessage(float, int, int);
     float temperature;
-    int sensorId;
-    int timestamp;
+    uint8_t sensorId;
+    uint8_t timestamp;
     SerialisedMessage serialise();
 };
